+++ database/migrations/2025_07_17_124458_create__commande_vetement_table.php
@@ -14,11 +14,11 @@ return new class extends Migration
         Schema::create('commande_vetement', function (Blueprint $table) {
             $table->id('commande_vetement_id');
 
-            $table->integer('commande_id');
+            $table->unsignedBigInteger('commande_id');
 
-            $table->integer('vetement_id');
+            $table->unsignedBigInteger('vetement_id');
 
-            $table->integer('quantite')->nullable();
+            $table->unsignedBigInteger('quantite')->nullable();
 
             $table->float('poids')->nullable();
